the checkFreeSpace() function overflowed for values greater than 2 gigabytes
git-svn-id: https://bostonbusmap.svn.sourceforge.net/svnroot/bostonbusmap@127 04fcf5f6-408b-428b-944a-e74021d52a6c

+++ bostonBusMap/src/boston/Bus/Map/database/DatabaseHelper.java
@@ -196,7 +196,7 @@ public class DatabaseHelper extends SQLiteOpenHelper
 		try
 		{
 			StatFs statFs = new StatFs(path);
-			int freeSpace = statFs.getAvailableBlocks() * statFs.getBlockSize(); 
+			long freeSpace = (long)statFs.getAvailableBlocks() * (long)statFs.getBlockSize(); 
 		
 			Log.v("BostonBusMap", "free database space: " + freeSpace);
 			return freeSpace >= 1024 * 1024 * 2;
